fix(rust): import_batch in multithread env
Batch import may cause the doc enter a temporary detached state. If
there are local edits happen during this phase, they may end up with
errors like 'AutoCommitNotStarted' or read-only errors.

This temporary phase should not be visible to users. This fix will lock
the transaction object during the temporary detached state so it's not
visible to another thread.